Turning 180 degrees coomand added
It is still not finished and always open for anyone to modify or who has
any sugestions. It is not quite finished yet, but I would like to have
it added into the main branch to let everyone know that a drive train
was added with an input from a gyro.